Extract prettyBytesBigint
